+++ packages/backends/src/node-file-trace.ts
@@ -1,18 +1,18 @@
 import {
   isBunVersion,
-  BuildOptions,
   FileBlob,
   FileFsRef,
-  Files,
+  type BuildOptions,
+  type Files,
+  type NodeVersion,
 } from '@vercel/build-utils';
 import { nodeFileTrace as nft } from '@vercel/nft';
 import { existsSync, lstatSync, readFileSync } from 'fs';
 import { join, relative } from 'path';
-import { type downloadInstallAndBundle } from './utils.js';
 
 export const nodeFileTrace = async (
   args: BuildOptions,
-  downloadResult: Awaited<ReturnType<typeof downloadInstallAndBundle>>,
+  nodeVersion: NodeVersion,
   output: {
     dir: string;
     handler: string;
@@ -21,7 +21,7 @@ export const nodeFileTrace = async (
   const { dir: outputDir, handler } = output;
   const entry = join(outputDir, handler);
   const files: Files = {};
-  const isBun = isBunVersion(downloadResult.nodeVersion);
+  const isBun = isBunVersion(nodeVersion);
   const conditions = isBun ? ['bun'] : undefined;
   const nftResult = await nft([entry], {
     base: args.repoRootPath,