+++ packages/build-utils/src/index.ts
@@ -35,7 +35,9 @@ import {
   getNodeBinPath,
   getNodeBinPaths,
   scanParentDirs,
+  findPackageJson,
   traverseUpDirectories,
+  PipInstallResult,
 } from './fs/run-user-scripts';
 import {
   getLatestNodeVersion,
@@ -84,6 +86,7 @@ export {
   runNpmInstall,
   runBundleInstall,
   runPipInstall,
+  PipInstallResult,
   runShellScript,
   runCustomInstallCommand,
   resetCustomInstallCommandSet,
@@ -102,6 +105,7 @@ export {
   isDirectory,
   getLambdaOptionsFromFunction,
   scanParentDirs,
+  findPackageJson,
   getIgnoreFilter,
   cloneEnv,
   hardLinkDir,
@@ -110,7 +114,7 @@ export {
 };
 
 export { EdgeFunction } from './edge-function';
-export { readConfigFile } from './fs/read-config-file';
+export { readConfigFile, getPackageJson } from './fs/read-config-file';
 export { normalizePath } from './fs/normalize-path';
 export { getOsRelease, getProvidedRuntime } from './os';
 
@@ -137,6 +141,9 @@ export {
   isExperimentalBackendsEnabled,
   isExperimentalBackendsWithoutIntrospectionEnabled,
   shouldUseExperimentalBackends,
+  PYTHON_FRAMEWORKS,
+  PythonFramework,
+  isPythonFramework,
 } from './framework-helpers';
 
 export * from './python';