+++ packages/build-utils/src/schemas.ts
@@ -25,6 +25,10 @@ const triggerEventSchema = {
       type: 'number',
       minimum: 0,
     },
+    maxConcurrency: {
+      type: 'number',
+      minimum: 1,
+    },
   },
   required: ['type', 'topic', 'consumer'],
   additionalProperties: false,