+++ packages/cli/test/mocks/client.ts
@@ -136,6 +136,14 @@ class MockStream extends PassThrough implements NodeJS.WriteStream {
   ref() {
     return this;
   }
+  destroySoon() {
+    return;
+  }
+  resetAndDestroy() {
+    return this;
+  }
+  autoSelectFamilyAttemptedAddresses: string[] = [];
+  pending = false;
   // END: Stub `WriteStream` interface to avoid TypeScript errors
 }
 
@@ -228,6 +236,7 @@ export class MockClient extends Client {
 
     output.initialize({
       stream: this.stderr,
+      supportsHyperlink: false,
     });
 
     this.argv = [];