+++ packages/introspection/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @vercel/introspection
 
+## 0.0.10
+
+### Patch Changes
+
+- Fix issue where introspection result could sometimes be split across stdout chunks ([#14638](https://github.com/vercel/vercel/pull/14638))
+
+- Disable network blocking logic during introspection while it's opt-in ([#14674](https://github.com/vercel/vercel/pull/14674))
+
+## 0.0.9
+
+### Patch Changes
+
+- - Parrallelize NFT and introspection steps ([#14619](https://github.com/vercel/vercel/pull/14619))
+  - Increase timeout to 8 seconds. It's taking up to 5 seconds for a large app we have been testing with
+  - Add more debug logs to introspection process
+
 ## 0.0.8
 
 ### Patch Changes