+++ packages/python-analysis/CHANGELOG.md
@@ -1,1 +1,13 @@
 # @vercel/python-analysis
+
+## 0.2.0
+
+### Minor Changes
+
+- use vitest instead of jest for testing ([#14681](https://github.com/vercel/vercel/pull/14681))
+
+## 0.1.1
+
+### Patch Changes
+
+- Fix release ([#14591](https://github.com/vercel/vercel/pull/14591))